feat(game): Implements crime and evidence system (#1396)
* feat(game): implement base of crime reporting and management system

- Introduce `CrimeManager` and `CrimeIdManager` to handle the lifecycle of in-game crime events, including database persistence and ID allocation.
- Implement `CSReportCrimePacket` logic to allow players to report crimes via evidence doodads (e.g., bloodstains).
- Add the `crime` database table and integrate it into the `SaveManager` loop.
- Provide a `doodad crime` admin subcommand for spawning test evidence with specific criminal and victim data.

* feat(game): implement bloodstain evidence generation for PvP crimes

- Generate small and large bloodstain doodads when damaging or killing friendly players.
- Refactor aggro management from `Npc` to the base `Unit` class to support player-to-player combat state tracking.
- Update `DoodadManager` to allow spawning system-driven doodads that bypass inventory requirements and housing checks.
- Implement `GetOwnerCharacter()` across various unit types to ensure crimes are correctly attributed to the responsible player.
- Add missing primary keys to the `crime` database table.

* feat(game): refactor CrimeManager to use dependency injection

- Replace static singleton access in `CrimeManager` with injected manager instances for `ICrimeIdManager`, `IZoneManager`, and `IDoodadManager`.
- Update `ICrimeManager` and `IDoodadManager` interfaces to expose methods required for evidence generation and system-driven doodad spawning.
- Added comments for placeholder logic of `DeletedEventIds` that supports upcoming trial and evidence tampering mechanics.

* feat(game): implement player assault tracking to improve crime reporting

- Track "AssaultedBy" and "AssaultOn" states to accurately identify the aggressor in PvP encounters.
- Prevent crime evidence (bloodstains) from being generated when a player retaliates against an initial attacker.
- Automatically clear assault records when combat state resets or upon character death.
- Refactor `CanAttack` so it is possible to attack friendly pets and vehicles when Bloodlust is enabled.

* feat(game): implement crime point accumulation and wanted status logic

- Apply the "Wanted" buff automatically when a character's crime points reach the defined threshold (50).
- Implement `SCCrimeChangedPacket` to synchronize crime and infamy point updates with the client.
- Add support for updating crime points for offline characters via direct database persistence.
- Refactor crime point values from `int` to `short` across managers, quests, and doodad functions to match protocol expectations.
- Update `SCCharacterStatePacket` to correctly report jury points.
- Refactor and expand `BuffConstants` to include "Wanted" status buff and updated some others.

* feat(game): implement theft evidence and pirate faction logic

- Generate footprint evidence when a player interacts with a doodad owned by another character.
- Implement Pirate faction transition, including automatic faction switching and expulsion from parties/expeditions when infamy thresholds are met.
- Prevent players from reporting their own crimes and log suspicious activity via `SusManager`.
- Update `SCCrimeChangedPacket` and faction broadcasting to improve client-side synchronization.
- Apply `Contemptuous` and `Wanted` buffs automatically based on crime and infamy records.

* feat(game): rename ZoneGroup to ZoneKey in crime tracking

- Rename the `zone_group` column in the `crime` database table to maintain naming consistency with the `ZoneManager` and underlying data structures and adjusted the related field names.

* fix(game): update crime database column name to zone_key

Forgot to change this one after changing the SQL
Also updated new crime related table creating SQL

+++ AAEmu.Game/Core/Managers/QuestManager.cs
@@ -1708,7 +1708,7 @@ public partial class QuestManager(ITaskManager taskManager, IZoneManager zoneMan
                     var parentComponent = GetComponentByActTemplate("QuestActSupplyCrimePoint", actId);
                     if (parentComponent == null)
                         continue;
-                    var template = new QuestActSupplyCrimePoint(parentComponent) { DetailId = actId, Point = reader.GetInt32("point") };
+                    var template = new QuestActSupplyCrimePoint(parentComponent) { DetailId = actId, Point = reader.GetInt16("point") };
                     AddActTemplate(template);
                 }
             }
